Added SqLite DB database to .gitignore
It wasn't added for several reasons:

It contains data that is specific to your local environment and is not usually needed by other developers or on production servers.
It can get large as more data is added, which can clutter your repository.
The database is often regenerated or updated automatically as part of the application setup.